Stamina largely removed
The stat is not actually removed. The enemies still use it to regulate how far they can go. However the players no longer need to concern themselves with their stamina because it has been increased to a number that means it will never run out during normal use. Because of that the stamina circle below the player has been removed.